+++ .eslintrc.js
@@ -1,9 +1,11 @@
 module.exports = {
-  ignorePatterns: ['__tests__/*', '__tests__/*.test.jsx'], // This will ignore all files in the __test__ folder
   settings: {
     react: {
       version: 'detect', // Automatically picks the version you have installed
     },
+  },
+  env: {
+    jest: true,  // Enables Jest global variables like 'test' and 'expect'
   },
   // Other ESLint configuration...
 };